+++ beacon_node/lighthouse_network/src/service/behaviour.rs
@@ -12,15 +12,10 @@ use libp2p::swarm::NetworkBehaviour;
 use libp2p::NetworkBehaviour;
 use types::EthSpec;
 
+use super::api_types::RequestId;
+
 pub type SubscriptionFilter = MaxCountSubscriptionFilter<WhitelistSubscriptionFilter>;
 pub type Gossipsub = BaseGossipsub<SnappyTransform, SubscriptionFilter>;
-
-/// Identifier of a request.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RequestId<AppReqId> {
-    Application(AppReqId),
-    Internal,
-}
 
 #[derive(NetworkBehaviour)]
 pub(crate) struct Behaviour<AppReqId: ReqId, TSpec: EthSpec> {
